+++ app/Http/Controllers/NoticiasController.php
@@ -4,14 +4,20 @@ namespace App\Http\Controllers;
 
 use App\Noticias;
 use Illuminate\Http\Request;
+use Auth;
 
 class NoticiasController extends Controller
 {
     public function indexAdmin()
     {
-        $noticias = Noticias::search()->orderBy('titulo_noticia')->paginate(20);
-    
-        return view('adminnoticias', compact('noticias'));
+        if(Auth::guard('admin')->check()){
+            $noticias = Noticias::search()->orderBy('titulo_noticia')->paginate(20);
+        
+            return view('adminnoticias', compact('noticias'));
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
     /**
@@ -43,28 +49,33 @@ class NoticiasController extends Controller
      */
     public function store(Request $request)
     {
-        //return $request->input('name');
-        //obliga a que los campos de name y email se llenen,
-        //si no no hace submit.
-        $this->validate($request, [
-            'titulo_noticia'=> 'required',
-            'descripcion_noticia'=> 'required',
-            'contenido_noticia'=> 'required',
+        if(Auth::guard('admin')->check()){
+            //return $request->input('name');
+            //obliga a que los campos de name y email se llenen,
+            //si no no hace submit.
+            $this->validate($request, [
+                'titulo_noticia'=> 'required',
+                'descripcion_noticia'=> 'required',
+                'contenido_noticia'=> 'required',
 
-        ]);
+            ]);
 
-        //  return 123;
-        //create new ReservaVisitaGuiada
-        $noticias = new Noticias;
-        $noticias->titulo_noticia = $request->input('titulo_noticia');
-        $noticias->descripcion_noticia = $request->input('descripcion_noticia');
-        $noticias->contenido_noticia = $request->input('contenido_noticia');
-        $noticias->fecha_noticia = now();
-        //save message
-        $noticias->save();
-    
-    
-        return redirect('/admin/noticias')->with('success','Noticia añadida con éxito.');
+            //  return 123;
+            //create new ReservaVisitaGuiada
+            $noticias = new Noticias;
+            $noticias->titulo_noticia = $request->input('titulo_noticia');
+            $noticias->descripcion_noticia = $request->input('descripcion_noticia');
+            $noticias->contenido_noticia = $request->input('contenido_noticia');
+            $noticias->fecha_noticia = now();
+            //save message
+            $noticias->save();
+        
+        
+            return redirect('/admin/noticias')->with('success','Noticia añadida con éxito.');
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
     /**
@@ -88,8 +99,13 @@ class NoticiasController extends Controller
      */
     public function edit($id)
     {
-        $noticia = Noticias::find($id);
-        return view('editarnoticia')->with('noticia', $noticia);
+        if(Auth::guard('admin')->check()){
+            $noticia = Noticias::find($id);
+            return view('editarnoticia')->with('noticia', $noticia);
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
     /**
@@ -101,26 +117,31 @@ class NoticiasController extends Controller
      */
     public function update(Request $request, $id)
     {
-        //return $request->input('name');
-        //obliga a que los campos de name y email se llenen,
-        //si no no hace submit.
-        $this->validate($request, [
-            'titulo_noticia'=> 'required',
-            'descripcion_noticia'=> 'required',
-            'contenido_noticia'=> 'required',
+        if(Auth::guard('admin')->check()){
+            //return $request->input('name');
+            //obliga a que los campos de name y email se llenen,
+            //si no no hace submit.
+            $this->validate($request, [
+                'titulo_noticia'=> 'required',
+                'descripcion_noticia'=> 'required',
+                'contenido_noticia'=> 'required',
 
-        ]);
+            ]);
+            
+            $noticias = Noticias::find($id);
+            $noticias->titulo_noticia = $request->input('titulo_noticia');
+            $noticias->descripcion_noticia = $request->input('descripcion_noticia');
+            $noticias->contenido_noticia = $request->input('contenido_noticia');
+
+            //save message
+            $noticias->save();
         
-        $noticias = Noticias::find($id);
-        $noticias->titulo_noticia = $request->input('titulo_noticia');
-        $noticias->descripcion_noticia = $request->input('descripcion_noticia');
-        $noticias->contenido_noticia = $request->input('contenido_noticia');
+        
+            return redirect('/admin/noticias')->with('success','Noticia modificada con éxito.');
+        }else{
 
-        //save message
-        $noticias->save();
-    
-    
-        return redirect('/admin/noticias')->with('success','Noticia modificada con éxito.');
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 
     /**
@@ -131,9 +152,14 @@ class NoticiasController extends Controller
      */
     public function destroy($id)
     {
-        $noticia = Noticias::find($id);
-        $noticia->delete();
-        
-        return redirect('/admin/noticias')->with('success', 'La noticia se ha eliminado correctamente.');
+        if(Auth::guard('admin')->check()){
+            $noticia = Noticias::find($id);
+            $noticia->delete();
+            
+            return redirect('/admin/noticias')->with('success', 'La noticia se ha eliminado correctamente.');
+        }else{
+
+            return redirect('/admin/login')->with('error','Debe estar conectado como administrador para entrar.');    
+        }
     }
 }
